requirements.txt
tensorflow>=2.13.0,<2.18.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
Pillow>=10.0.0
imageio>=2.31.0
openpyxl>=3.1.0
pydot>=1.4.2
graphviz>=0.20.1
